IControl slave parameters tabs can now be deleted Tab Header equals thBe type of the plugin this should closes #390
TextBoxHidden is now implemented in WSM parameters. Databinding is impossible for PasswordBoxes, because of security issues. This closes #400

RadioButtons are now bound through DataBinding to the equal enum in the Settings. Every RadioButton has his own DataBinding to the enum. The problem of reactions between multiple RadioButtons and the single enum were solved though a converter. this will close #98