Removed dynamic properties from BooleanFunctionParser (refs #402)
Please note: this seems to be the only plugin which suffers a functional degradation with static data properties.
However, I can't really estimate the impact as there is no template which uses the input props (there's an IControl one but this is unaffected by my change).
Without a template I also cannot test whether everything's still working.
If multiple input props are needed, I suggest to add more static properties and implement a "hide property mechanism" which should allow for the advantages of displaying a variable number of inputs but without the complexity (leading to various failures) of dynamic properties.